main.py
from dotenv import load_dotenv
from langchain_mistralai import ChatMistralAI
from langchain_community.document_loaders import TextLoader
from langchain_core.prompts import ChatPromptTemplate
load_dotenv()
data = TextLoader("documents_loader/data.txt")
docs = data.load()
model = ChatMistralAI(model = "mistral-small-2506")
template =ChatPromptTemplate.from_messages(
    [("system","You are an AI that summarises the text"),
     ("human","{data}")])

prompt = template.format_messages(data = docs[0].page_content)

result = model.invoke(prompt)
print(result.content)