pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-mistralai>=1.1.2",
    "langchain-openai>=1.1.11",
    "lxml>=6.0.2",
    "mistralai>=2.0.5",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "unstructured>=0.21.5",
    "uvicorn>=0.42.0",
]
